feat: add basic crud function

+++ src/pages/dashboard/index.tsx
@@ -1,10 +1,12 @@
 import { useState, useCallback } from 'react';
 import { NextPage } from 'next';
 import { useSelector, useDispatch } from 'react-redux';
+import swal from 'sweetalert';
 
 import Layout from '../../components/layout/Layout';
 import { Modal, Table } from '../../components/UI';
-import { selectUser } from '../../infrastructure/store/users/userSlice';
+import { DataType } from '../../types';
+// import { selectUser } from '../../infrastructure/store/users/userSlice';
 
 import {
   syncIssue,
@@ -13,7 +15,7 @@ import {
 } from '../../infrastructure/store/data/dataSlice';
 
 const Dashboard: NextPage = () => {
-  const [filterArr, setFilterArr] = useState([]);
+  const [filterArr, setFilterArr] = useState<DataType[]>([]);
   const [isModalOpen, setIsModalOpen] = useState({
     isOpen: false,
     isEdit: false,
@@ -24,8 +26,14 @@ const Dashboard: NextPage = () => {
   const dispatch = useDispatch();
 
   const syncIssueHandler = useCallback(() => {
-    dispatch(syncIssue());
-    alert('Sync');
+    swal({
+      text: 'Are you sure that you want to sync data?',
+      dangerMode: false,
+    }).then((willSync) => {
+      if (willSync) {
+        dispatch(syncIssue());
+      }
+    });
   }, []);
 
   const createIssueHandler = useCallback(() => {
@@ -47,7 +55,16 @@ const Dashboard: NextPage = () => {
   );
 
   const deleteIssueHandler = useCallback((id) => {
-    dispatch(deleteIssue(id));
+    swal({
+      title: 'Are you sure?',
+      text: 'Are you sure that you want to delete this issue?',
+      icon: 'warning',
+      dangerMode: true,
+    }).then((willDelete) => {
+      if (willDelete) {
+        dispatch(deleteIssue(id));
+      }
+    });
   }, []);
 
   return (
